refactor(内容读取): 收拢工具编排与内部服务边界
- 将内容读取的三个 Tool facade 和共用参数 schema 收归 tool_content_read/tools
- 将加载、窗口构建、读取模型与 readers 收归 tool_content_read/services
- 删除并列的 tool_content_read_tools 包，容器和测试统一改用领域包内的新路径
