+++ services/wisepen-chat-service/src/chat/application/tools/session_tools/tool_content_read/__init__.py
@@ -1,4 +1,4 @@
-from .models import (
+from .services.models import (
     ToolContentReadFailure,
     ToolContentRegexMatch,
     ToolContentRegexReadRequest,
