+++ services/wisepen-chat-service/src/chat/application/tools/session_tools/tool_content_read/services/__init__.py
@@ -1,0 +1,1 @@
+"""内容读取工具的内部服务实现。"""
