+++ services/wisepen-chat-service/src/chat/application/tools/session_tools/tool_content_read/services/content_loader.py
