+++ services/wisepen-chat-service/src/chat/application/tools/session_tools/tool_content_read/services/readers/sequential_reader.py
