+++ services/wisepen-chat-service/src/chat/container.py
@@ -38,7 +38,7 @@ from chat.application.tools.skill_tools.utils.skill_matcher import DefaultSkillM
 from chat.application.tools.skill_tools import LoadSkillAssetTool
 from chat.application.tools.skill_tools import LoadSkillTool
 from chat.application.tools.core import ToolRegistry
-from chat.application.tools.session_tools.tool_content_read_tools import (
+from chat.application.tools.session_tools.tool_content_read.tools import (
     ToolContentRegexReadTool,
     ToolContentRankedExpandReadTool,
     ToolContentSequentialReadTool,
